When laterFiles indicates that the sentinel file has been lost, restart with an initial collection
When we run out of newerFiles, eventually newer files will appear and get processed. When the entire data directory(/-ies) are cleared, however, we lose the sentinel file, and it's unlikely that the particular value will reappear again. Therefore, we should treat the disappearance of the sentinel as a restart, and do a full initial collection, also considering the --initial-{first,last} arguments.