onbag: WIP: Counting of changes does not work; fix for polling
Unlike in the original onchange, executeTrigger() is invoked inside a pipeline here, so the observedChanges value won't persist across invocations. (Even worse, when using onchange, each change will trigger a separate instance of onbag, but we'll get to that later.)
The counting needs to move to the first process within the pipeline; as a different approach is needed for onchange, anyway, move it to pollForOutput() instead of processBag().