bag: Reimplementation: Replace tee with read loop
The problem with tee is that when lines are popped off the top, updating from a slow source (like interactive terminal input) stops. As we don't need high throughput here and don't expect to handle huge amounts of data, an explicit loop should be fine.
The existing bag is truncated at the beginning, before first input arrives, to match the behavior of tee.