onbag: Improve handling of multiple consumed lines with --stop-on PATTERN
Discarding the whole block when there's a stop pattern in it makes the behavior hard to predict. We'd better still process any lines before the stop pattern, and put those after it back into the bag (but at the end, as the bag doesn't yet provide an interface for resubmitting at the beginning).
The extended algorithm also fixes a problem with parallel execution, where the read line(s) would be silently dropped if the marker file vanished. Those lines need to be put back as well.