tempdir: ENH: Allow sole determination (+ creation) of temp dir via --print
For some clients, the passing of commands and intricate logic may be infeasible, and setting up a recursive launch of itself also is not straightforward. For these cases, just leverage the directory determination logic, at the cost of having to the chdir, error handling, and possible cleanup (which is further complicated if --or-existing is used - then the best the client can possibly do is let a created temp directory linger and hope that the operating system will eventually clean this up).