onbag: FIX: When onchange does polling, we don't necessarily get all bag updates
By default, only single lines from the bag are consumed, but several lines may be put into the bag at once. When inotifywait tracks the bag, that still seems to (mostly?) work, but now that the --stop-after TIMEOUT may be passed to onchange and it then has to resort to polling, only the first of added lines is processed.
In the onbag --onchange handler, we need to invoke processBag() until is fails, to ensure that every added line in the bag has been processed.