addOrUpdateBlock: CHG: --add-after to --add-before
We'll never catch and update a block that was inserted on a previous invocation if we're appending after ADDRESS; as soon as we match ADDRESS, further searching is aborted and stuff gets inserted (again and again). What works for the single-line insertion in addOrUpdateLine cannot be applied here.
Instead, do the updating at ADDRESS (before any existing text), so that it will be before ADDRESS on the next invocation, and then gets found and updated.