addOrUpdateBlock: Adding before empty line not properly handled
The parameter --add-before '^$' is a corner case, because the resulting line is empty.
Need to do the ADDRESS match check before clearing the current line to avoid an accidental match.
The check for an existing block forgot to swap back hold and pattern space; usually, this is inconsequential, because any output would be treated as "update happened", but with an empty line, this is interpreted as "no update yet", and the block that already got updated gets appended once more.