onbag: Consume existing lines at the beginning
Not just when changes to the bag have been observed. This is the expected behavior, and it also addresses the inconsistency in behavior between the polling and the onchange implementations: Whereas polling would (slowly) detect and consume existing lines, the onchange algorithm would only start the consumption once at least one external change is observed. (But then the removals of the lines would essentially feed themselves.)