onbag: Omit the constant count=0 when not counting, and read the count from the database before processing the bag
When not counting, the $observedChanges variable isn't updated; including it as a constant 0 in the prompt is of no use.
When counting, the counter is only updated after the trigger, which results in some invocations having them, and then the next not. By reading the existing count prior to invoking processBag(), we always have the correct count included.