onbag: Indicate the last COMMAND's exit status in the following prompt also when consuming the existing lines
This so far only worked for the observed lines, as runWithPrompt has a trailing prompt there. For the existing lines, the previous COMMAND's exit status need to be forwarded to the next one's, using runWithPrompt's new --initial-status option, through the runWithPromptPromptArgs array (renamed from runWithPromptTrailingPromptArg). But first of all, that status needs to find its way back to consumeExistingOutput(), which we can do by returning it (a variable won't do, as multiple pipelines are involved in between). The only complication there is that the COMMAND's exit status must not masquerade as one of the signals.