[on]bag[-consume]: ENH: Correctly handle --{pre,ap}pend-{before,after}-once when using --parallel
Parallel execution triggers each execution in a job, so the clearing of the *OnceText variables has no effect. We need to use an external flag through executionMarker (transactional for concurrency control) to get this across, using a separate database for each main script invocation (through $$ - each job uses $BASHPID as $$ still is the main script's), and a single entry for all variables (so that only one job wins and can update each variable that is set).